refactor: move save_response to SDK
